Wish list

+++ src/main/java/com/seven/marketclip/wishList/service/WishListsService.java
@@ -6,6 +6,7 @@ import com.seven.marketclip.exception.CustomException;
 import com.seven.marketclip.exception.ResponseCode;
 import com.seven.marketclip.goods.domain.Goods;
 import com.seven.marketclip.goods.repository.GoodsRepository;
+import com.seven.marketclip.security.UserDetailsImpl;
 import com.seven.marketclip.wishList.domain.WishLists;
 import com.seven.marketclip.wishList.repository.WishListsRepository;
 import lombok.RequiredArgsConstructor;
@@ -13,8 +14,7 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 
-import static com.seven.marketclip.exception.ResponseCode.GOODS_NOT_FOUND;
-import static com.seven.marketclip.exception.ResponseCode.SUCCESS;
+import static com.seven.marketclip.exception.ResponseCode.*;
 
 @Service
 @RequiredArgsConstructor
@@ -23,19 +23,32 @@ public class WishListsService {
     private final WishListsRepository wishListsRepository;
 
     @Transactional
-    public ResponseCode doWishList(Long goodsId, Account account) throws CustomException {
+    public ResponseCode doWishList(Long goodsId, UserDetailsImpl account, String httpMethod) throws CustomException {
+        Account detailsAccount = new Account(account);
         Goods goods = goodsRepository.findById(goodsId).orElseThrow(
                 () -> new CustomException(GOODS_NOT_FOUND)
         );
-        WishLists wishLists = wishListsRepository.findByGoodsAndAccount(goods, account).orElse(null);
-        if(wishLists != null){
-            wishListsRepository.delete(wishLists);
-        }else {
-            wishListsRepository.save(WishLists.builder()
-                    .goods(goods)
-                    .account(account)
-                    .build());
+        WishLists wishLists = wishListsRepository.findByGoodsAndAccount(goods, detailsAccount).orElse(null);
+        if (wishLists != null) {
+            if (httpMethod.equals("DELETE")) {
+                wishListsRepository.delete(wishLists);
+            } else {
+                throw new CustomException(WRONG_WISHLIST_SAVE_REQUEST);
+            }
+        } else {
+            if (httpMethod.equals("POST")) {
+                wishListsRepository.save(WishLists.builder()
+                        .goods(goods)
+                        .account(detailsAccount)
+                        .build());
+            } else {
+                throw new CustomException(WRONG_WISHLIST_DELETE_REQUEST);
+            }
         }
         return SUCCESS;
+    }
+
+    public int wishListCount(Long goodsId) {
+        return wishListsRepository.findAllByGoodsId(goodsId).size();
     }
 }
